pyproject.toml
[project]
name = "alice-bp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.3",
    "django-split-settings>=1.3.2",
    "djangorestframework>=3.16.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "dj-database-url>=2.2.0",
    "django-filter>=24.2",
]

[tool.setuptools]
packages = ["alice_skill", "config", "pyanywhere_bg", "tgbot_bp"]

[dependency-groups]
dev = [
    "pytest-django>=4.11.1",
    "pytest-aiohttp>=1.0.0",
]

[tool.uv.workspace]
members = [
    "tgbot_bp",
]
[tool.ruff.format]
quote-style = "single"
